[include]: cast last to increase precision when performing math
git-svn-id: https://svn.parrot.org/parrot/trunk@20404 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ include/parrot/thr_windows.h
@@ -65,7 +65,7 @@ typedef HANDLE Parrot_thread;
        { \
          ++(c).m_lWaiters; \
          UNLOCK(m); \
-         diff = ((DWORD)t->tv_sec - (DWORD)sec)*1000L + (t->tv_nsec - nsec)/1000000L; \
+         diff = (DWORD)((t->tv_sec - sec)*1000L + (t->tv_nsec - nsec)/1000000L); \
          WaitForSingleObject((c).m_hSema, diff); \
          LOCK(m); \
          --(c).m_lWaiters; \
